ocr/requirements.txt
appwrite
pytesseract
Pillow
requests
